+++ Classes/CoverPreferences.h
@@ -1,0 +1,17 @@
+//
+//  CoverPreferences.h
+//  VGTagger
+//
+//  Created by Bilal Syed Hussain on 10/08/2011.
+//  Copyright 2011 St. Andrews KY16 9XW. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+
+
+@interface CoverPreferences :  NSViewController <MASPreferencesViewController> {
+@private
+    
+}
+
+@end