+++ Classes/CoverPreferences.m
@@ -1,0 +1,17 @@
+
+//
+//  CoverPreferences.m
+//  VGTagger
+//
+//  Created by Bilal Syed Hussain on 10/08/2011.
+//  Copyright 2011 St. Andrews KY16 9XW. All rights reserved.
+//
+
+#import "CoverPreferences.h"
+
+
+@implementation CoverPreferences
+
+
+
+@end